lots of work
Now my functions in src don't print to stdout/stderr except in ifdef'ed
regions that aren't called when package is built for distribution.
Fukuda's code has the possibility of making such calls.  Most but not
quite all of his fprintf's are guarded by if(localdebug) or something
of the sort and cannot be called when the package is built for distribution,
but not all of them.  So this is a bug but cannot be fixed without going
over Fukuda's code and changing some of it.

Some of the calls to rr_* and rrf_* functions were wrong ones, now fixed.
Header file mycddio_f.h fixed to be like mycddio.h

Still to do: have to actually fix makeH, which was why this version was
started.

Also added .Rinstignore